refactor, apply google c++ recommended class names
